Revert "move runToolWithStreaming in @lib/api (#15493)"
This reverts commit 1fc789099cdfc028efa7c1f7ade79e22dccb0579.
