+++ front/lib/resources/agent_mcp_action_resource.ts
@@ -28,7 +28,7 @@ import { ConversationResource } from "@app/lib/resources/conversation_resource";
 import { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
 import type { ReadonlyAttributesType } from "@app/lib/resources/storage/types";
 import type { ModelStaticWorkspaceAware } from "@app/lib/resources/storage/wrappers/workspace_models";
-import { getResourceIdFromSId, makeSId } from "@app/lib/resources/string_ids";
+import { makeSId } from "@app/lib/resources/string_ids";
 import type { ResourceFindOptions } from "@app/lib/resources/types";
 import logger from "@app/logger/logger";
 import type { ModelId, Result } from "@app/types";
@@ -137,24 +137,18 @@ export class AgentMCPActionResource extends BaseResource<AgentMCPActionModel> {
     });
   }
 
-  static async fetchById(
+  static async fetchByModelIdWithAuth(
     auth: Authenticator,
-    sId: string,
+    id: ModelId,
     transaction?: Transaction
   ): Promise<AgentMCPActionResource | null> {
-    const modelId: ModelId | null = getResourceIdFromSId(sId);
-    if (!modelId) {
-      return null;
-    }
-
     const [action] = await this.baseFetch(
       auth,
       {
-        where: { id: modelId },
+        where: { id },
       },
       transaction
     );
-
     return action;
   }
 
@@ -169,14 +163,6 @@ export class AgentMCPActionResource extends BaseResource<AgentMCPActionModel> {
         },
       },
     });
-  }
-
-  static async fetchByModelIdWithAuth(
-    auth: Authenticator,
-    id: ModelId
-  ): Promise<AgentMCPActionResource | null> {
-    const [action] = await this.fetchByModelIds(auth, [id]);
-    return action ?? null;
   }
 
   static async listBlockedActionsForConversation(
